Move vector lowering to its own file
