+++ mlir/lib/Gradient/Transforms/CMakeLists.txt
@@ -6,6 +6,7 @@ file(GLOB SRC
     gradient_bufferize.cpp
     LoweringPatterns.cpp
     gradient_lowering.cpp
+    vector_lowering.cpp
     ConversionPatterns.cpp
     gradient_to_llvm.cpp
 )
